// src/app/models/product.ts
export class product {
  id:any;
  name: any | undefined;
  user_id: any | undefined;
  category_id: any | undefined;
  brand_id: any | undefined;
  seller_id: any | undefined;
  thumbnail_image: any | undefined;
  image: any []=[];
  tags: any[] = [];
  short_description: any | undefined;
  description: any | undefined;
  price: any | undefined;
  quantity: any | undefined;
  status: any | undefined;
  feature_product: any | undefined;
  slug: any | undefined;
}