// src/app/models/blog.ts
export class blog {
  id: any;
  name: any | undefined;
  blog_category_id: any | undefined;
  image: any | undefined;
  short_description: any | undefined;
  description: any | undefined;
  slug: any | undefined;
  status: any | undefined;
  top_blog: any | undefined;
  created_at: any | undefined;
}